refactor: modularize spa
