+++ src/app.js
@@ -12,11 +12,11 @@ app.use(express.json());
 
 app.use(express.static(path.join(__dirname, "public")));
 
+app.use("/api", routes);
+
 app.get("*", (req, res) => {
   res.status(200).sendFile(path.join(__dirname, "public", "index.html"));
 });
-
-app.use("/api", routes);
 
 app.listen(port, () => {
   console.log(`Server running on "http://${hostname}:${port}"`);
