+++ src/public/index.html
@@ -3,14 +3,10 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>EasyBarber SPA</title>
-    <script type="module" src="./components/components.js"></script>
+    <title>EasyBarber</title>
+    <script type="module" src="./js/index.js"></script>
 </head>
 <body>
-    <div id="app"></div>
-    <script type="module">
-        import "./components/components.js";
-    </script>
-    
+    <div id="root"></div>
 </body>
 </html>
