+++ src/public/js/event.js
@@ -1,0 +1,6 @@
+export default function onNavigate(path) {
+    return new CustomEvent("onstatechange", {
+        detail: { path },
+    });
+    }
+
